refactoring modules under only try submodule
