+++ lnspec_py/messages_types/tvl_record.py
@@ -1,8 +1,7 @@
 from curses import raw
-from ..fundamental.ints import bigsizeInt
 
 """
-This class represent a single TVL record 
+This class represent a single TVL record
 as specify in https://github.com/lightning/bolts/blob/master/01-messaging.md#type-length-value-format
 """
