--- tests/test_lnspec_py.py
@@ -1,5 +1,0 @@
-from lnspec_py import __version__
-
-
-def test_version():
-    assert __version__ == "0.1.0"
