+++ tests/test_tvl_record.py
@@ -1,6 +1,6 @@
 import pytest
 
-from lnspec_py.abstract.tvl_record import TVLRecord
+from lnspec_py.messages_types.tvl_record import TVLRecord
 
 """
 Test for TVL reader
